fix: move component folder
