+++ app/components/faulty/styles.js
