+++ app/components/loading/styles.js
