+++ app/components/timer/Timer.js
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 
 import { View, Text } from "react-native";
 
-import { msToTime } from "../../../store/services/helpers/moment";
+import { msToTime } from "../../store/services/helpers/moment";
 import styles from "./styles";
 
 class Timer extends Component {
