+++ app/components/timer/styles.js
